+++ frame/plasm-staking/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "pallet-plasm-session"
+name = "pallet-plasm-staking"
 version = "0.1.0"
 authors = ["Takumi Yamashita <takumi@stake.co.jp>"]
 edition = "2018"
@@ -70,11 +70,6 @@ git = 'https://github.com/staketechnologies/substrate.git'
 branch = 'plasm-master'
 package = 'pallet-contracts'
 
-[dependencies.validator-manager]
-default_features = false
-path = "../validator-manager"
-package = "pallet-validator-manager"
-
 [dependencies.sp-staking]
 default_features = false
 git = 'https://github.com/staketechnologies/substrate.git'
@@ -88,6 +83,11 @@ version = '1.0'
 [dependencies.serde_derive]
 optional = true
 version = '1.0'
+
+[dependencies.validator-manager]
+default-features = false
+path = "../validator-manager"
+package = "pallet-validator-manager"
 
 [dev-dependencies]
 wabt = "0.9.2"
@@ -108,9 +108,8 @@ std = [
     'session/std',
     'timestamp/std',
     'system/std',
-    'validator-manager/std',
     'contract/std',
-    "sp-staking/std",
+    'sp-staking/std',
    	'serde_derive',
 	'serde/std',
 ]
