--- frame/validator-manager/src/traits.rs
@@ -1,6 +1,0 @@
-pub trait OnEraEnding<ValidatorId, EraIndex> {
-	fn on_era_ending(
-		_ending: EraIndex,
-		_start_era: EraIndex,
-	) -> Option<Vec<ValidatorId>>;
-}
